refactor: moved event card into event card screen and reworked api calls
