+++ packages/app/screens/group/event_overview/EventOverviewScreen.js
@@ -1,13 +1,10 @@
-import React, { useEffect, useState } from "react";
+import React from "react";
 import { View, StyleSheet, Image } from "react-native";
 import PropTypes from "prop-types";
-import axios from "axios";
 
 import EventOverviewText from "@app/components/EventOverview/EventOverviewText";
 import EventOverviewRegister from "@app/components/EventOverview/EventOverviewRegister";
 import FontAwesome5 from "@expo/vector-icons/FontAwesome5";
-import { ENDPOINT } from "@app/utils/constants";
-import * as SecureStore from "@app/utils/SecureStore";
 
 const WHITE = "white";
 const transparentColor = "#00000077";
@@ -47,42 +44,8 @@ const styles = StyleSheet.create({
 });
 
 export default function EventOverviewScreen({ navigation, route }) {
-  const { eventID, previousScreen } = route.params;
-  const [event, setEvent] = useState([]);
-
-  const getEvents = async () => {
-    const token = await SecureStore.getValueFor("accessToken");
-    const { data: response } = await axios.get(
-      `${ENDPOINT}/events/${eventID}`,
-      {
-        headers: {
-          Authorization: token,
-        },
-      }
-    );
-    setEvent(response.data.event);
-  };
-
-  useEffect(() => {
-    getEvents();
-  }, []);
-
-  function formatDate(rawDate) {
-    const date = new Date(rawDate);
-    let hours = date.getUTCHours();
-    let minutes = date.getMinutes();
-    let dayPeriod = "AM";
-
-    if (hours > 12) {
-      hours -= 12;
-      dayPeriod = "PM";
-    }
-    if (minutes < 10) minutes = "0" + minutes;
-
-    return `${date.getMonth() + 1}/${date.getDate()}/${
-      date.getFullYear() - 2000
-    } at ${hours}:${minutes} ${dayPeriod}`;
-  }
+  const { previousScreen, title, timeBegin, timeEnd, location, description } =
+    route.params;
 
   const onBackPress = () => {
     navigation.navigate(previousScreen);
@@ -111,11 +74,11 @@ export default function EventOverviewScreen({ navigation, route }) {
       <View style={styles.overview}>
         <EventOverviewText
           groupName="Software Engineering Association"
-          title={event.title}
-          timeBegin={formatDate(event.startDate)}
-          timeEnd={formatDate(event.endDate)}
-          location={event.location}
-          description={event.description}
+          title={title}
+          timeBegin={timeBegin}
+          timeEnd={timeEnd}
+          location={location}
+          description={description}
         />
       </View>
       <View style={styles.register}>
